Adds 72 Character Auto Wrap
Using the `width` property of CSS, this will auto wrap messages to 72
characters. Because CSS does not know about character width, only pixel
width, the formula of 1 character to 8.00000000000007 pixels was used.
This results to our magic number of 576 pixels to 72 characters.

Of course, you can remove / change this as needed. Perhaps 78 or even 80
characters? Its up to you.

In addition, you may instead add a background image which can be
provided as a guide to wrapping text -- more useful.

Enjoy!

Signed-off-by: Adam Brenner <aebrenne@uci.edu>

Conflicts:
	gmail-fixed.css